% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreg_mean_sd.R
\name{aggreg_mean_sd}
\alias{aggreg_mean_sd}
\title{Compute the mean and standard deviation of the given numeric vector}
\usage{
aggreg_mean_sd(x, ...)
}
\arguments{
\item{x}{the vector to be processed}

\item{...}{further arguments passed to \code{\link[=format_number]{format_number()}}}
}
\value{
The formatted string containing the mean and standard deviation.
}
\description{
The function returns the formatted \code{"mean ± stdev"}
of vector \code{x}.
}
\seealso{
\code{\link[=aggreg_median_iqr]{aggreg_median_iqr()}}
}
\author{
Michal Burda
}
