% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreg_median.R
\name{aggreg_median}
\alias{aggreg_median}
\title{Compute the median of the given numeric vector}
\usage{
aggreg_median(data, var, ...)
}
\arguments{
\item{data}{the data frame}

\item{var}{the name of the column to be processed}

\item{...}{further arguments passed to \code{\link[=format_number]{format_number()}}}
}
\value{
The formatted string containing the median.
}
\description{
The function returns the formatted "median"
of column \code{var} in data frame \code{data}.
}
\seealso{
\code{\link[=aggreg_median_iqr]{aggreg_median_iqr()}}
}
\author{
Michal Burda
}
