% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreg_median_iqr.R
\name{aggreg_median_iqr}
\alias{aggreg_median_iqr}
\title{Compute the median and lower/upper quartile of the given numeric vector}
\usage{
aggreg_median_iqr(x, ...)
}
\arguments{
\item{x}{the vector to be processed}

\item{...}{further arguments passed to \code{\link[=format_number]{format_number()}}}
}
\value{
The formatted string containing the median and lower/upper quartile.
}
\description{
The function returns the formatted "median (lower_quartile -- upper_quartile)"
of vector \code{x}.
}
\seealso{
\code{\link[=aggreg_mean_sd]{aggreg_mean_sd()}}
}
\author{
Michal Burda
}
