% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreg_percent.R
\name{aggreg_percent}
\alias{aggreg_percent}
\title{Compute the relative frequency (percentage) of the given level in the given vector}
\usage{
aggreg_percent(x, level = 1, use_na = FALSE, ...)
}
\arguments{
\item{x}{the vector to compute the occurences of \code{level} in. It must be an atomic vector.}

\item{level}{the level to be computed. It could be either integerish value or a string or \code{NA}.
The integerish value is treated as the index of level to be counted, character value selects
the level to be counted by its name, and \code{NA} selects missing values to be counted.}

\item{use_na}{if \code{FALSE} then \code{NA} values are omitted prior computing the frequency.}

\item{...}{further arguments passed to \code{\link[=format_percent]{format_percent()}}}
}
\value{
The relative number of occurences of \code{level} in \code{x}, i.e., the fraction in \link{0,1}.
}
\description{
The function returns the relative number (percentage) of occurences of \code{level} in vector \code{x}.
}
\examples{
aggreg_percent(c(F,F,T,T,T), level = 1)                          # 3/5
aggreg_percent(c(F,F,T,T,T), level = "FALSE")                    # 2/5
aggreg_percent(c("b", "a", "a"), level = "a")                    # 2/3
aggreg_percent(c(NA, "b", "a", "a"), level = NA, use_na = TRUE)  # 1/4
}
\seealso{
\code{\link[=aggreg_count]{aggreg_count()}}
}
\author{
Michal Burda
}
