% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat0.R
\name{cat0}
\alias{cat0}
\title{Output the objects with default separator being empty string.}
\usage{
cat0(..., sep = "")
}
\arguments{
\item{...}{the objects to output}

\item{sep}{The separator}
}
\description{
Output the objects with default separator being empty string.
}
\seealso{
\code{\link[=cat]{cat()}}
}
\author{
Michal Burda
}
