% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_numeric_aggreg.R
\name{create_numeric_aggreg}
\alias{create_numeric_aggreg}
\title{Factory function for creation of aggregation function of numeric vectors}
\usage{
create_numeric_aggreg(
  type = c("robust", "parametric", "test"),
  parametric_aggreg = aggreg_mean_sd,
  robust_aggreg = aggreg_median_iqr,
  normality_test = shapiro.test,
  normality_thresh = 0.05
)
}
\arguments{
\item{type}{selection of the statistic to compute for numeric
variables: \code{"robust"} (see \code{\link[=aggreg_median_iqr]{aggreg_median_iqr()}}), \code{"parametric"} (see
\code{\link[=aggreg_mean_sd]{aggreg_mean_sd()}}), or \code{"test"}, which performs \code{.normality_test} first
on each numeric variable to determine between \code{"robust"} and \code{"parametric"}}

\item{parametric_aggreg}{which aggregator to use as a parametric one}

\item{robust_aggreg}{which aggregator to use as a robust one}

\item{normality_test}{a function that performs the normality test in case of
\code{type = "test"}. The function must return a list with the \code{"p.value"}
element containing the p-value of the rejection of normality.}

\item{normality_thresh}{a p-value threshold below which will be the
\code{normality_test} interpreted as rejection of normality.}
}
\value{
An aggregator function
}
\description{
Function creates a function for aggregating numeric vectors into a single
character value. It creates three types of aggregators: robust, parametric,
or test-based accordingly to the \code{type} argument.
}
\examples{
aggreg <- create_numeric_aggreg("robust")
aggreg(1:10)

}
\seealso{
\code{\link[=aggreg_mean_sd]{aggreg_mean_sd()}}, \code{\link[=aggreg_median_iqr]{aggreg_median_iqr()}}
}
\author{
Michal Burda
}
