% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_numeric_test.R
\name{create_numeric_test}
\alias{create_numeric_test}
\title{Factory function for creation of test function for numeric vectors}
\usage{
create_numeric_test(
  type = c("robust", "parametric", "test"),
  k = 2,
  parametric_test = if (k > 2) test_aov else test_t,
  robust_test = if (k > 2) test_kruskal else test_wilcox,
  normality_test = shapiro.test,
  normality_thresh = 0.05
)
}
\arguments{
\item{type}{selection of the type of statistical test for numeric
variables: \code{"robust"} (see \code{\link[=test_wilcox]{test_wilcox()}} or \code{\link[=test_kruskal]{test_kruskal()}}),
\code{"parametric"} (see \code{\link[=test_t]{test_t()}} or \code{\link[=test_aov]{test_aov()}}), or \code{"test"}, which
performs \code{.normality_test} first on the numeric variable to select from
\code{"robust"} and \code{"parametric"}}

\item{k}{the number of groups for which the test will be performed. The value
must be greater than 1. If \code{k} equals \code{2}, two-sample tests are returned
(either \code{\link[=test_t]{test_t()}} or \code{\link[=test_wilcox]{test_wilcox()}}). For \code{k} greater than \code{2}, either
\code{\link[=test_aov]{test_aov()}} or \code{\link[=test_kruskal]{test_kruskal()}} is returned.}

\item{parametric_test}{which test to use as a parametric one}

\item{robust_test}{which test to use as a robust one}

\item{normality_test}{a function that performs the normality test in case of
\code{type = "test"}. The function must return a list with the \code{"p.value"}
element containing the p-value of the rejection of normality.}

\item{normality_thresh}{a p-value threshold below which will be the
\code{normality_test} interpreted as rejection of normality.}
}
\value{
A test function

a test function
}
\description{
Function creates a function for statistical testing of numeric vectors
by groups. It creates three types of test functions: robust, parametric,
or test-based accordingly to the \code{type} argument.
}
\seealso{
\code{\link[=test_t]{test_t()}}, \code{\link[=test_wilcox]{test_wilcox()}}, \code{\link[=test_kruskal]{test_kruskal()}}, \code{\link[=test_aov]{test_aov()}}
}
\author{
Michal Burda
}
