% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_call.R
\name{do_call}
\alias{do_call}
\title{Call function with arguments as a list while possibly ignoring unused arguments.}
\usage{
do_call(
  fun,
  ...,
  .args = list(),
  .ignore_unused = TRUE,
  .quote = FALSE,
  .envir = parent.frame()
)
}
\arguments{
\item{fun}{the function to be called}

\item{...}{further arguments to be passed to \code{fun}}

\item{.args}{the named list of arguments}

\item{.ignore_unused}{if TRUE, ignore arguments in \code{args} that are not accepted by \code{fun}.
If \code{fun} accepts \code{...} arguments, nothing is ignored.}

\item{.quote}{a logical value indicating whether to quote the arguments.}

\item{.envir}{an environment within which to evaluate the call. This will be most useful
if what is a character string and the arguments are symbols or quoted expressions.}
}
\value{
The result of the function \code{fun}
}
\description{
Call function with arguments as a list while possibly ignoring unused arguments.
}
\seealso{
\code{\link[base:do.call]{base::do.call()}}
}
\author{
Michal Burda
}
