% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreach_col_by_class.R
\name{foreach_col_by_class}
\alias{foreach_col_by_class}
\title{Apply a function selected by class for each column of data frame}
\usage{
foreach_col_by_class(.data, .handlers = list(), .fail = NULL, ...)
}
\arguments{
\item{.data}{the data frame to be processed}

\item{.handlers}{the named list of callback functions (the names have to
correspond to \code{.data} column classes)}

\item{.fail}{the value returned if the callback function is not found}

\item{...}{further arguments passed to the callback function}
}
\value{
The list of results of function calls or \code{.fail} if the proper
callback function cannot be found in \code{.hanlders}.
}
\description{
For each column of \code{.data}, an appropriate callback function is selected
from \code{.handlers} and executed. Selection is based on column class and
names of the elements of \code{.handlers}: first element (function) in \code{.handlers},
that has a name as some class of the processed column, will be executed.
If search for the callback function fails, the value of \code{.fail} will be
used as a result.
}
\examples{
df <- data.frame(a = factor(c("a", "a", "b", "b")),
                 b = 1:4,
                 c = 11:14)
foreach_col_by_class(df, list(factor = nlevels,
                              integer = sum))
# returns list(a = 2, c = 50)
}
\author{
Michal Burda
}
