% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_bars.R
\name{format_bars}
\alias{format_bars}
\title{Convert the data into strings reporting relative frequencies in the form of
textual bar-chart}
\usage{
format_bars(p, n, width = 20, symbol = "I", fill = "-")
}
\arguments{
\item{p}{an integerish vector of positive cases.}

\item{n}{an integerish vector of all cases.}

\item{width}{the maximum width of the textual bar-chart (i.e., the maximum number
of \code{symbol} repeats)}

\item{symbol}{the character string used to plot the bar}

\item{fill}{the character string used to fill the empty part of the chart}
}
\value{
a character vector of bars created from \code{symbol} that depict the
relative frequency of \code{p}s in \code{n}s
}
\description{
\code{p} is the number of positive cases and \code{n} is the number of all cases. The result
is a character vector depicting bars created from the \code{symbol} (repeated at most
\code{width}-times if \code{p} = \code{n}).
}
\author{
Michal Burda
}
