% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_count_percent.R
\name{format_count_percent}
\alias{format_count_percent}
\title{Convert the data into strings reporting the absolute and relative frequencies}
\usage{
format_count_percent(p, n, ignore_na = FALSE, ...)
}
\arguments{
\item{p}{a numeric vector of positive cases.}

\item{n}{a numeric vector of all cases.}

\item{ignore_na}{if \code{FALSE}, the \code{NA} values will result into something like \code{"NA (NA \%)"}.
If \code{TRUE}, such results will be replaced with an empty string. Any other value
will be directly used as a replacement.}

\item{...}{further arguments passed to \code{\link[=format_percent]{format_percent()}}}
}
\value{
a character vector of size \code{length(p)} (if \code{p} and \code{n} is specified)
}
\description{
\code{p} is the number of positive cases and \code{n} is the number of all cases. The result
is a character vector formatted as \code{"<p> (<p/n> \%)"}.
}
\author{
Michal Burda
}
