% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_enumeration.R
\name{format_enumeration}
\alias{format_enumeration}
\title{Convert the character vector to a string of enumerated values.}
\usage{
format_enumeration(x, sep = ", ", max = 5, dots = "...", ...)
}
\arguments{
\item{x}{the character vector to be converted}

\item{sep}{the separator string that will be placed between the \code{x} vector values}

\item{max}{the maximum number of values from \code{x} to be placed into the string}

\item{dots}{the string appended to the result if the \code{x} vector is larger than \code{max}}

\item{...}{further arguments that are ignored}
}
\value{
The string with concatenated values from \code{x}
}
\description{
Convert the character vector to a string of enumerated values.
}
\author{
Michal Burda
}
