% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_interval.R
\name{format_interval}
\alias{format_interval}
\title{Convert two numeric vectors to strings formatted as intervals}
\usage{
format_interval(from, to = NULL, sep = " -- ", ...)
}
\arguments{
\item{from}{the first vector of values. If \code{to} is \code{NULL}, the length of \code{from}
must be 2.}

\item{to}{the second vector of values. If given, the length of \code{to} must be equal
to the length of \code{from}.}

\item{sep}{the separator string}

\item{...}{further arguments passed to \code{\link[=format_number]{format_number()}}}
}
\value{
A string vector with elements formatted as: \verb{"from -- to}
}
\description{
Convert two numeric vectors to strings formatted as intervals
}
\author{
Michal Burda
}
