% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_number.R
\name{format_number}
\alias{format_number}
\title{Convert the numeric vector to character vector}
\usage{
format_number(x, digits = 2, na = "NA", style = "signif", ...)
}
\arguments{
\item{x}{the numeric vector to be converted}

\item{digits}{the number of decimal point numbers to round the numbers to}

\item{na}{the string to be used for \code{NA} values}

\item{...}{further arguments that are ignored}
}
\value{
a character vector of transformed numeric values
}
\description{
Convert the numeric vector to character vector
}
\author{
Michal Burda
}
