% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_percent.R
\name{format_percent}
\alias{format_percent}
\title{Convert the numeric vector to a character vector in a percentage formatting.}
\usage{
format_percent(x, ...)
}
\arguments{
\item{x}{a numeric vector to be transformed}

\item{...}{further arguments passed to \code{\link[=format_number]{format_number()}}}
}
\value{
a character vector
}
\description{
The numeric vector is multiplied by 100, rouned, and the percent sign (\verb{\%}) is appended to it.
}
\author{
Michal Burda
}
