% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_string.R
\name{format_string}
\alias{format_string}
\title{Shorten string and add \code{"..."}}
\usage{
format_string(x, max_width = Inf, ellipsis = "...", ...)
}
\arguments{
\item{x}{the character vector to be formatted}

\item{max_width}{the maximum width of strings in \code{x}}

\item{ellipsis}{string that will be appended to trimmed elements of \code{x}}

\item{...}{unused arguments}
}
\value{
a formatted character vector
}
\description{
The elements of character vector \code{x} are trimmed to the maximum length
given by \code{max_width}. If shortening takes place, \code{ellipsis} is appended
to the end of the shortened string.
}
\author{
Michal Burda
}
