% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_plot.R
\name{freq_plot}
\alias{freq_plot}
\title{Plot the frequency bar chart with absolute and relative count in text over each
bar.}
\usage{
freq_plot(x, name = deparse(substitute(x)))
}
\arguments{
\item{x}{The factor data to plot}

\item{name}{the name of the data axis}
}
\description{
Plot the frequency bar chart with absolute and relative count in text over each
bar.
}
\author{
Michal Burda
}
