% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram}
\alias{histogram}
\title{Plot a nice histogram}
\usage{
histogram(
  x,
  name = deparse(substitute(x)),
  rugs = length(x) <= 1000,
  na = any(is.na(x))
)
}
\arguments{
\item{x}{numeric vector}

\item{name}{the name to put on the X-axis}

\item{rugs}{whether to draw rugs}

\item{na}{whether to draw the NA bar}
}
\description{
Plot a nice histogram
}
\author{
Michal Burda
}
