% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_outlier.R
\name{is_outlier}
\alias{is_outlier}
\title{Test whether the values are outliers}
\usage{
is_outlier(x, extreme = FALSE)
}
\arguments{
\item{x}{the numeric vector to be tested}

\item{extreme}{\code{TRUE} if test for extreme outliers}
}
\value{
a logical vector
}
\description{
Test whether the values are outliers
}
\author{
Michal Burda
}
