% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modus.R
\name{modus}
\alias{modus}
\title{Get the most frequent value}
\usage{
modus(x, na_rm = FALSE)
}
\arguments{
\item{x}{the vector within which to search the modus}

\item{na_rm}{whether to ignore \code{NA}}
}
\value{
The modus (mode), i.e., the most frequent value
}
\description{
Get the most frequent value
}
\author{
Michal Burda
}
