% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVsumstats.R
\name{MVsumstats}
\alias{MVsumstats}
\title{MVsumstats function}
\usage{
MVsumstats(by, bx, sy, sx, N, cor.x = NULL, cor.xy = NULL)
}
\arguments{
\item{by}{beta coefficients for the outcome (a J-vector for J genetic variants)}

\item{bx}{beta coefficients for the exposures (a J x K matrix, where K is the number of exposures, and the k-th column are the J-vector of beta coefficients for the k-th exposure)}

\item{sy}{standard errors for the outcome}

\item{sx}{standard errors for the exposures (a J x K matrix, where the k-th column are the J-vector of standard errors for the k-th exposure)}

\item{N}{sample size; a one-sample analysis is performed if only one sample size is entered. If two sample sizes N=c(Ny,Nx) are entered, then a two-sample analysis is performed under the assumption that genetic associations with traits with the outcome (Ny-sample) and exposures (Nx-sample) are measured from non-overlapping samples}

\item{cor.x}{(optional) K x K exposure correlation matrix (where K is the number of exposures); default is a diagonal matrix so the exposures are assumed to be unconditionally uncorrelated (results are not too sensitive to this choice)}

\item{cor.xy}{(optional) K x 1 exposure-outcome correlation vector (where K is the number of exposures); default is a vector of zeros, so that the outcome and exposures are unconditionally uncorrelated (again, results are not too sensitive to this choice)}

\item{ld}{genetic variant (LD) correlation matrix}
}
\value{
Output is a list containing:
\itemize{
\item \code{GamY} \cr
Vector of regression coefficients from regressing the outcome on the joint J-vector of instruments
\item \code{GamX} \cr
Matrix of regression coefficients where the k-th column corresponds to regressing the k-th exposure on the joint J-vector of instruments
\item \code{SigY} \cr
a J-vector of variances, where the j-th element corresponds to the variance of the j-th element of GamY
\item \code{SigX} \cr
a J-list of K x K matrices covariance matrices where the j-th list corresponds to the variance-covariance matrix for the j-th row of GamX
\item \code{SigXY} \cr
a J-list of K-vectors, where the j-th element corresponds to the covariance of the j-th row of GamX with j-th element of GamY
}
}
\description{
A method for transforming one-at-a-time uncorrelated univariable regression summary statistics into multivariate regression summary statistics
}
\details{
This method transforms one-at-a-time uncorrelated univariable regression summary statistics into multivariate regression summary statistics for the joint vector of instrument associations with the exposure and outcome.
}
\author{
Jack Bowden, Stephen Burgess, Ashish Patel, Dmitry Shungin
}
