% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_if_fails.R
\name{na_if_fails}
\alias{na_if_fails}
\title{Return NA if expression fails}
\usage{
na_if_fails(expr, na = NA_real_)
}
\arguments{
\item{expr}{an expression to evaluate}

\item{na}{value to return if the expression fails (default is \code{NA_real_})}
}
\value{
The result of the expression if it succeeds, or \code{na} if it fails
}
\description{
This function evaluates an expression and returns \code{NA} if it fails, i.e.,
if it stops with error.
}
\author{
Michal Burda
}
