% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_row.R
\name{new_row<-}
\alias{new_row<-}
\title{Append new rows to a data frame}
\usage{
new_row(data) <- value
}
\arguments{
\item{data}{a data frame to append the rows to}

\item{value}{new data to be apended to \code{data}}
}
\description{
Append new rows to a data frame
}
