% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCsumstats.R
\name{PCsumstats}
\alias{PCsumstats}
\title{PCsumstats function}
\usage{
PCsumstats(by, sy, bx, sx, ld, N, pc.thres)
}
\arguments{
\item{by}{beta coefficients for the outcome (a J-vector for J genetic variants)}

\item{sy}{standard errors for the outcome}

\item{bx}{beta coefficients for the exposures (a J x K matrix, where K is the number of exposures, and the k-th column are the J-vector of beta coefficients for the k-th exposure)}

\item{sx}{standard errors for the exposures (a J x K matrix, where the k-th column are the J-vector of standard errors for the k-th exposure)}

\item{ld}{genetic variant (LD) correlation matrix}

\item{N}{K+1 vector of sample sizes; the first element is the sample size for outcome associations. The (k+1)-th element is the sample size for the k-th exposure associations}

\item{pc.thres}{(optional) the proportion of variance in genetic associations that the principal components of genetic variants used as instruments are required to explain; default is 0.99 so that at least 99\% of variation in the genetic variant LD matrix will be explained by the instruments}
}
\value{
Output is a list containing:
\itemize{
\item \code{by} \cr
Vector of outcome beta-coefficients for the the principal components of genetic variants
\item \code{sy} \cr
Vector of outcome standard errors for the the principal components of genetic variants
\item \code{bx} \cr
Matrix of exposure beta-coefficients for the the principal components of genetic variants (the k-th column corresponds to the k-th exposure)
\item \code{sx} \cr
Matrix of exposure standard errors for the the principal components of genetic variants (the k-th column corresponds to the k-th exposure)
\item \code{pc.thres} \cr
the proportion of variance in genetic associations that the principal components of genetic variants used as instruments explain
}
}
\description{
A method for calculating summary association data using principal components of many highly correlated variants
}
\details{
This method transforms univariable genetic associations based a set of many highly correlated genetic variants into univariable genetic associations based on a smaller set of uncorrelated principal components of the many correlated variants. \cr
}
\author{
Jack Bowden, Stephen Burgess, Ashish Patel, Dmitry Shungin
}
