% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_frequency.R
\name{plot_frequency}
\alias{plot_frequency}
\title{Plot the frequency bar chart with absolute and relative count in text over each
bar.}
\usage{
plot_frequency(x, name = deparse(substitute(x)))
}
\arguments{
\item{x}{The factor data to plot}

\item{name}{the name of the data axis}
}
\description{
Plot the frequency bar chart with absolute and relative count in text over each
bar.
}
\author{
Michal Burda
}
