% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune}
\alias{prune}
\title{Pruning function}
\usage{
prune(by, sy, bx, sx, ld, prune.thres)
}
\arguments{
\item{by}{beta coefficients for the outcome (a J-vector for J genetic variants)}

\item{sy}{standard errors for the outcome}

\item{bx}{beta coefficients for the exposures (a J x K matrix, where K is the number of exposures, and the k-th column are the J-vector of beta coefficients for the k-th exposure)}

\item{sx}{standard errors for the exposures (a J x K matrix, where the k-th column are the J-vector of standard errors for the k-th exposure)}

\item{ld}{genetic variant (LD) correlation matrix}

\item{prune.thres}{(optional) R^2 pruning threshold for variant correlations; if unspecified the default value is R^2 = 0.95}
}
\value{
Output is a list containing:
\itemize{
\item \code{by} \cr
Vector of outcome beta-coefficients for the genetic variants retained after pruning
\item \code{sy} \cr
Vector of outcome standard errors for the genetic variants retained after pruning
\item \code{bx} \cr
Matrix of exposure beta-coefficients for the genetic variants retained after pruning (the k-th column corresponds to the k-th exposure)
\item \code{sx} \cr
Matrix of exposure standard errors for the genetic variants retained after pruning (the k-th column corresponds to the k-th exposure)
\item \code{ld} \cr
matrix of genetic variant (LD) correlations for the genetic variants retained after pruning
\item \code{prune.thres} \cr
the R^2 pruning threshold used
}
}
\description{
Selects relevant variants for all exposures in a way that maximises collective instrument strength for all exposures in a multivariable model
}
\details{
The method leaves a set of genetic variants that are collectively relevant for all exposures in the model. \cr
}
\author{
Jack Bowden, Stephen Burgess, Ashish Patel, Dmitry Shungin
}
