% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q1.R
\name{q1}
\alias{q1}
\title{Compute the first quartile (25\%)}
\usage{
q1(x, na_rm = FALSE)
}
\arguments{
\item{x}{the numeric vector}

\item{na_rm}{whether to ignore \code{NA}}
}
\value{
The first quartile
}
\description{
Compute the first quartile (25\%)
}
\author{
Michal Burda
}
