% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q3.R
\name{q3}
\alias{q3}
\title{Compute the third quartile (75\%)}
\usage{
q3(x, na_rm = FALSE)
}
\arguments{
\item{x}{the numeric vector}

\item{na_rm}{whether to ignore \code{NA}}
}
\value{
The third quartile
}
\description{
Compute the third quartile (75\%)
}
\author{
Michal Burda
}
