% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plot.R
\name{qq_plot}
\alias{qq_plot}
\title{Plot the quantile-quantile plot}
\usage{
qq_plot(x)
}
\arguments{
\item{x}{the numeric data to plot}
}
\value{
the plot
}
\description{
Plot the quantile-quantile plot
}
\author{
Michal Burda
}
