% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrange.R
\name{rearrange}
\alias{rearrange}
\title{Rearrange the vector of values by the names in other vector}
\usage{
rearrange(value, by)
}
\arguments{
\item{value}{the named vector whose values have to be rearranged}

\item{by}{the rearangment specification: a character vector of
names of \code{value}}
}
\value{
a new vector of the same type as \code{value}
}
\description{
Create a vector of length equal to \code{by} filled with \code{NA} values.
Then put a value of \code{value} to corresponding places such that
the value of \code{by} equals to the name of \code{value}.
}
\examples{
by <- c("a", "b", "b", "c", "a")
values <- c(a = 1, b = 2)
rearrange(values, by)       # returns c(1, 2, 2, NA, 1)
}
