% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_outliers.R
\name{report_outliers}
\alias{report_outliers}
\title{Generate a summary table of outliers in numeric columns of a data frame}
\usage{
report_outliers(.data)
}
\arguments{
\item{.data}{the data frame}
}
\value{
a summary of outliers as a tibble
}
\description{
Generate a summary table of outliers in numeric columns of a data frame
}
\author{
Michal Burda
}
