% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_overview.R
\name{report_overview}
\alias{report_overview}
\title{Generate a data frame overview table}
\usage{
report_overview(.data, ...)
}
\arguments{
\item{.data}{a data frame}

\item{...}{other agruments passed to \code{\link[=format_enumeration]{format_enumeration()}},
\code{\link[=format_bars]{format_bars()}} or \code{\link[=format_count_percent]{format_count_percent()}}}
}
\value{
an overview table as a tibble
}
\description{
Generate a data frame overview table
}
\author{
Michal Burda
}
