% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_aov.R
\name{test_aov}
\alias{test_aov}
\title{Return p-value of the Analysis of Variance}
\usage{
test_aov(x, g, confounders = NULL, ...)
}
\arguments{
\item{x}{numeric values}

\item{g}{grouping factor}

\item{confounders}{\code{NULL} or a list of confounders, i.e. a list of numeric
vectors of the same lengths as \code{x} and \code{g}}

\item{...}{further parameters passed to the underlying \code{\link[=aov]{aov()}} function}
}
\value{
The p-value of the test
}
\description{
This function is a wrapper over the \code{\link[=aov]{aov()}} function.
It performs the analysis of variance and returns the resulting p-value.
}
\author{
Michal Burda
}
