% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_chisq.R
\name{test_chisq}
\alias{test_chisq}
\title{Return p-value of the Chi-squared test}
\usage{
test_chisq(x, g, ...)
}
\arguments{
\item{x}{data factor}

\item{g}{grouping factor}

\item{...}{further parameters passed to the underlying \code{\link[=chisq.test]{chisq.test()}} function}
}
\value{
The p-value of the test
}
\description{
This function is a wrapper over the \code{\link[=chisq.test]{chisq.test()}} function.
It performs Chi-squared goodness-of-fit test and returns the resulting p-value.
}
\author{
Michal Burda
}
