% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fisher.R
\name{test_fisher}
\alias{test_fisher}
\title{Return p-value of Fisher's test}
\usage{
test_fisher(x, g, ...)
}
\arguments{
\item{x}{data factor}

\item{g}{grouping factor}

\item{...}{further parameters passed to the underlying \code{\link[=fisher.test]{fisher.test()}} function}
}
\value{
The p-value of the test
}
\description{
This function is a wrapper over the \code{\link[=fisher.test]{fisher.test()}} function.
It performs Fisher's test and returns the resulting p-value.
}
\author{
Michal Burda
}
