% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_kruskal.R
\name{test_kruskal}
\alias{test_kruskal}
\title{Return p-value of Kruskal-Wallis rank sum test}
\usage{
test_kruskal(x, g, ...)
}
\arguments{
\item{x}{numeric values}

\item{g}{grouping factor}

\item{...}{further parameters passed to the underlying \code{\link[=kruskal.test]{kruskal.test()}} function}
}
\value{
The p-value of the test
}
\description{
This function is a wrapper over the \code{\link[=kruskal.test]{kruskal.test()}} function.
It performs the Kruskal-Wallis rank sum test and returns the resulting p-value.
}
\author{
Michal Burda
}
