% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_wilcox.R
\name{test_wilcox}
\alias{test_wilcox}
\title{Return p-value of the two-sample Wilcoxon rank sum test}
\usage{
test_wilcox(x, g, ...)
}
\arguments{
\item{x}{numeric values}

\item{g}{grouping factor (must contain 2 levels)}

\item{...}{further parameters passed to the underlying \code{\link[=wilcox.test]{wilcox.test()}} function}
}
\value{
The p-value of the test
}
\description{
This function is a wrapper over the \code{\link[=wilcox.test]{wilcox.test()}} function.
It performs the two-sample Wilcoxon rank sum test and returns
the resulting p-value.
}
\author{
Michal Burda
}
