% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist_first.R
\name{unlist_first}
\alias{unlist_first}
\title{Flatten list by taking only its first element.}
\usage{
unlist_first(x)
}
\arguments{
\item{x}{the list to be simplified}
}
\value{
a vector
}
\description{
Given a list \code{x}, \code{unlist_first} simplifies it to produce a vector
which contains all first values of each element of \code{x}.
}
\author{
Michal Burda
}
