% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreg_count.R
\name{aggreg_count}
\alias{aggreg_count}
\title{Compute the count of given level in given vector}
\usage{
aggreg_count(x, level = 1)
}
\arguments{
\item{x}{the vector to compute the occurences of \code{level} in. It must be an atomic vector.}

\item{level}{the level to be computed. It could be either integerish value or a string or \code{NA}.
The integerish value is treated as the index of level to be counted, character value selects
the level to be counted by its name, and \code{NA} selects missing values to be counted.}
}
\value{
The number of occurences of \code{level} in \code{x}
}
\description{
The function returns the number of occurrences of \code{level} in vector \code{x}.
}
\examples{
aggreg_count(c(F,F,T,T,T), level = 1)           # 3
aggreg_count(c(F,F,T,T,T), level = "FALSE")     # 2
aggreg_count(c("b", "a", "a"), level = "a")     # 2
aggreg_count(c(NA, "b", "a", "a"), level = NA)  # 1
}
\author{
Michal Burda
}
