% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreg_mean.R
\name{aggreg_mean}
\alias{aggreg_mean}
\title{Compute the mean of the given numeric vector}
\usage{
aggreg_mean(data, var, ...)
}
\arguments{
\item{data}{the data frame}

\item{var}{the name of the column to be processed}

\item{...}{further arguments passed to \code{\link[=format_number]{format_number()}}}
}
\value{
The formatted string containing the mean.
}
\description{
The function returns the formatted \code{"mean"}
of column \code{var} in data frame \code{data}.
}
\seealso{
\code{\link[=aggreg_mean_sd]{aggreg_mean_sd()}}
}
\author{
Michal Burda
}
