% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreg_median.R
\name{aggreg_median}
\alias{aggreg_median}
\title{Compute the median of the given numeric vector}
\usage{
aggreg_median(x, ...)
}
\arguments{
\item{x}{the vector to be processed}

\item{...}{further arguments passed to \code{\link[=format_number]{format_number()}}}
}
\value{
The formatted string containing the median.
}
\description{
The function returns the formatted "median"
of vector \code{x}.
}
\seealso{
\code{\link[=aggreg_median_iqr]{aggreg_median_iqr()}}
}
\author{
Michal Burda
}
