% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreg_sd.R
\name{aggreg_sd}
\alias{aggreg_sd}
\title{Compute the standard deviation of the given numeric vector}
\usage{
aggreg_sd(data, var, ...)
}
\arguments{
\item{data}{the data frame}

\item{var}{the name of the column to be processed}

\item{...}{further arguments passed to \code{\link[=format_number]{format_number()}}}
}
\value{
The formatted string containing the standard deviation.
}
\description{
The function returns the formatted \code{"stdev"}
of column \code{var} in data frame \code{data}.
}
\seealso{
\code{\link[=aggreg_median_iqr]{aggreg_median_iqr()}}
}
\author{
Michal Burda
}
