% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{baseline}
\alias{baseline}
\title{Create baseline table from data frame}
\usage{
baseline(
  .data,
  .n = TRUE,
  .all = TRUE,
  .test = n_groups(.data) > 0,
  .bullet = " • ",
  .numeric_aggreg = create_numeric_aggreg("robust"),
  .categ_aggreg = aggreg_count_percent,
  ...
)
}
\arguments{
\item{.data}{the data frame to be processed}

\item{.n}{if \code{TRUE}, the first row of the result has to contain number of rows
in groups}

\item{.all}{if \code{TRUE}, an additional \code{"all"} column is created containing
statistics of the whole data frame}

\item{.test}{if \code{TRUE}, an additional \code{"p"} column is created with p-values
of appropriate statistical tests of significance of differences in groups}

\item{.bullet}{a string used as a bullet in enumeration of factor levels}

\item{.numeric_aggreg}{a custom function used to aggregate numeric values.}

\item{.categ_aggreg}{a custom function used to aggregate categorical values.}
}
\description{
Function works mainly for data frames grouped by \code{\link[dplyr:group_by]{dplyr::group_by()}}. In the
resulting data frame, each column corresponds to a group and each row
corresponds to a non-grouping column of the original data frame.
}
\seealso{
\code{\link[=create_numeric_aggreg]{create_numeric_aggreg()}}
}
\author{
Michal Burda
}
