% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pvalue.R
\name{format_pvalue}
\alias{format_pvalue}
\title{Convert the numeric vector of p-values to formatted character vector}
\usage{
format_pvalue(
  x,
  digits = 4,
  thresh = 0.1,
  varname = "",
  ns = "NS",
  na = "NA",
  adjust = NULL,
  ...
)
}
\arguments{
\item{x}{the numeric vector to be converted}

\item{digits}{the number of decimal point numbers to round the numbers to}

\item{thresh}{a value greater than this threshold is considered as
non-significant}

\item{varname}{if \code{""} then simply formatted p-values are returned. If
\code{varname} contains some string, then the result is formatted as
\code{"varname = value"}}

\item{ns}{the string used for non-significant values}

\item{na}{the string to be used for \code{NA} values}

\item{adjust}{if not \code{NULL}, the p-values are adjusted using the method
specified in this argument (see \code{\link[=p.adjust]{p.adjust()}} for possible methods)}

\item{...}{further arguments that are ignored}
}
\value{
a character vector of transformed numeric values
}
\description{
P-values in \code{x} are formatted as strings with given number of \code{digits}.
If the p-value is greater than the given threshold \code{thresh}, the \code{"NS"}
string is returned instead (or whathever is set in the \code{ns} argument).
If the p-value is lower than the lowest number that fits to the given
number of \code{digits}, a string similar to\code{"<0.0001"} (having the appropriate
number of digits) is returned. If the p-value is \code{NA}, the \code{"NA"} string
is returned (or whatever is set in the \code{na} argument).
}
\details{
If \code{varname} argument is not an empty string, the p-values are formatted
as \code{"<varname>=<value>"}.
}
\author{
Michal Burda
}
