% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pvalue.R
\name{format_pvalue}
\alias{format_pvalue}
\title{Convert the numeric vector of p-values to formatted character vector}
\usage{
format_pvalue(x, digits = 4, thresh = 0.05, varname = "", ns = "NS", na = "NA")
}
\arguments{
\item{x}{the numeric vector to be converted}

\item{digits}{the number of decimal point numbers to round the numbers to}

\item{na}{the string to be used for \code{NA} values}
}
\value{
a character vector of transformed numeric values
}
\description{
P-values in \code{x} are formatted as strings with given number of \code{digits}.
If the p-value is greater than the given threshold \code{thresh}, the \code{"NS"}
string is returned instead (or whathever is set in the \code{ns} argument).
If the p-value is lower than the lowest number that fits to the given
number of \code{digits}, a string similar to\code{"<0.0001"} (having the appropriate
number of digits) is returned. If the p-value is \code{NA}, the \code{"NA"} string
is returned (or whatever is set in the \code{na} argument).
}
\details{
If \code{varname} argument is not an empty string, the p-values are formatted
as \code{"<varname>=<value>"}.
}
\author{
Michal Burda
}
