% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammer-package.R
\docType{package}
\name{hammer-package}
\alias{hammer}
\alias{hammer-package}
\title{hammer: Personal package of miscellaneous R functions for data analyses}
\description{
Personal package of miscellaneous R functions for data analyses
}
\author{
\strong{Maintainer}: Michal Burda \email{michal.burda@osu.cz} (\href{https://orcid.org/0000-0002-4182-4407}{ORCID})

}
\keyword{internal}
