% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammer-package.R
\docType{package}
\name{hammer-package}
\alias{hammer}
\alias{hammer-package}
\title{hammer: Monte-Carlo Cross-Validation with Weights}
\description{
An experimental variant of Monte-Carlo cross-validation which uses case weights to determine the probability of a given observation being selected for the analysis set. The key function in the package is a minor extension of the 'mc_cv()` function from 'rsample'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mikemahoney218/hammer}
  \item \url{https://mikemahoney218.github.io/hammer/}
  \item Report bugs at \url{https://github.com/mikemahoney218/hammer/issues}
}

}
\author{
\strong{Maintainer}: Michael Mahoney \email{mike.mahoney.218@gmail.com} (\href{https://orcid.org/0000-0003-2402-304X}{ORCID})

}
\keyword{internal}
