% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hammer.R
\name{hammer}
\alias{hammer}
\title{Heterogeneity-Adjusted Multivariable MEndelian Randomization}
\usage{
hammer(
  by,
  bx,
  sy,
  sx,
  N,
  alpha = 0.05,
  ld = NULL,
  cor.x = NULL,
  cor.xy = NULL,
  robust = TRUE,
  pc.thres = NULL,
  clump.thres = NULL,
  exposure.names = NULL
)
}
\arguments{
\item{by}{beta coefficients for the outcome (a J-vector for J genetic variants)}

\item{bx}{beta coefficients for the exposures (a J x K matrix, where K is the number of exposures, and the k-th column are the J-vector of beta coefficients for the k-th exposure)}

\item{sy}{standard errors for the outcome}

\item{sx}{standard errors for the exposures (a J x K matrix, where the k-th column are the J-vector of standard errors for the k-th exposure)}

\item{N}{sample size; a one-sample analysis is performed if only one sample size is entered. If two sample sizes N=c(Ny,Nx) are entered, then a two-sample analysis is performed under the assumption that genetic associations with traits with the outcome (Ny-sample) and exposures (Nx-sample) are measured from non-overlapping samples}

\item{alpha}{(optional) (1-alpha)*100\% is the nominal coverage of confidence intervals reported in the table of resuls; default is 0.05}

\item{ld}{genetic variant (LD) correlation matrix}

\item{cor.x}{(optional) K x K exposure correlation matrix (where K is the number of exposures); default is a diagonal matrix so the exposures are assumed to be unconditionally uncorrelated (results are not too sensitive to this choice)}

\item{cor.xy}{(optional) K x 1 exposure-outcome correlation vector (where K is the number of exposures); default is a vector of zeros, so that the outcome and exposures are unconditionally uncorrelated (again, results are not too sensitive to this choice)}

\item{robust}{(optional) logical: if \code{TRUE} then the results are adjusted for overdispersion heterogeneity in genetic variant associations with the outcome if such heterogeneity is detected; default is \code{TRUE}}

\item{pc.thres}{(optional) the proportion of variance in genetic associations that the principal components of genetic variants used as instruments are required to explain; default is 0.99 so that at least 99\% of variation in genetic variants will be explained by the instruments}

\item{clump.thres}{(optional) R^2 clumping threshold for variant correlations; if unspecified the default value is R^2 = 0.95}

\item{exposure.names}{(optional) a J x 1 character vector of the names of exposures included in the model}
}
\value{
Output is a list containing:
\itemize{
\item \code{est} \cr
K x 1 vector of exposure causal effect estimates on the outcome
\item \code{se} \cr
K x 1 vector of standard errors relating to the exposure causal effect estimates
\item \code{condF} \cr
K x 1 vector of conditional F-statistics (only reported if number of exposures is K > 1), otherwise a scalar F-statistic in the univariable exposure-outcome model is reported
\item \code{kappa} \cr
the estimated overdispersion heterogeneity parameter (a parameter estimate > 0 means the model was adjusted for overdispersion heterogeneity); only reported if the input \code{robust=TRUE}
\item \code{numberIVs} \cr
the number of instruments used by the method
}
}
\description{
A method for multivariable cis-Mendelian randomization analysis with summary data.
}
\details{
This method uses the principal components of genetic variants in a single gene region to instrument multiple exposures for Mendelian randomization randomization analyses.
}
\author{
Jack Bowden, Stephen Burgess, Ashish Patel, Dmitry Shungin
}
