% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_columns.R
\name{report_columns}
\alias{report_columns}
\title{Generate a table describing all columns in a data frame}
\usage{
report_columns(.data)
}
\arguments{
\item{.data}{a data frame}
}
\value{
a description of columns as tibble
}
\description{
Generate a table describing all columns in a data frame
}
\author{
Michal Burda
}
